Colocando fonte e separando o js do slide em outro arquivo
