+++ src/data.js
@@ -1,4 +1,6 @@
-// estas funciones son de ejemplo
+//manipulação de dados através de arrays e objetos
+
+//Exemplo filterData(data, condition)
 
 export const example = () => {
   return 'example';
@@ -6,4 +8,18 @@ export const example = () => {
 
 export const anotherExample = () => {
   return 'OMG';
+};
+
+//data é onde estão os dados!! função pode ter 3 parâmetros
+export const ordemAlfabetica = (data) =>{
+
+  let ordemAZ = data.sort(function(a, z){
+    if (a.name < z.name){
+      return ;
+    } 
+    else {
+      return ;
+    }
+  })
+  return ordemAZ
 };
