+++ src/index.html
@@ -9,6 +9,7 @@
   </head>
   <body class="background">
     <header>
+      <div><img src="https://occ-0-2794-2219.1.nflxso.net/dnm/api/v6/LmEnxtiAuzezXBjYXPuDgfZ4zZQ/AAAABQOxGMxzPfr3lvtrDxZ9ahcGCw4wMQsdBljUXHiuon2Oa1SmjgGWYx-QWkyLuFwEaguoI3yaq592H-XkHzLbtgh5Pu9ljDJV6AzS.png?r=1ab" alt="" class="box-logo"></div>
       <div class="box">
         <div class="box-image"><img src="img/hate.gif" /></div>
         <div class="box-image"><img src="img/clones.gif" /></div>
@@ -19,10 +20,16 @@
     </header>
 
     <main>
-     
+      <form>
+        <button class="btn-az" id="btn-az">A-Z</button>
+        <button class="btn-za" id="btn-za">Z-A</button>
       <section class="search">
         <input type="search" placeholder="Digite a sua pesquisa.." />
       </section>
+
+      
+    </form>
+
     </main>
 
     <footer>
@@ -51,7 +58,7 @@
         /></a>
       </section>
     </footer>
-
+    <script src="slide.js" type="module"></script>
     <script src="main.js" type="module"></script>
   </body>
 </html>
