+++ src/style.css
@@ -1,4 +1,7 @@
-@import url("https://fonts.googleapis.com/css2?family=Big+Shoulders+Stencil+Text:wght@100&family=Inter:wght@600&family=Mali:ital,wght@1,700&family=Mochiy+Pop+P+One&family=Roboto:ital,wght@1,500&family=Yanone+Kaffeesatz:wght@200;700&display=swap");
+@font-face {
+  font-family: "Shlop";
+  src: url("font/shlop\ rg.ttf") format("truetype");
+}
 
 * {
   padding: 0;
@@ -7,23 +10,37 @@
   box-sizing: border-box;
 }
 
+:root {
+  --body-font: "Share Tech", sans-serif;
+  --green-neon: #5fe431;
+  --light-green: #85d97aa8;
+  --light-blue: #11aebf;
+}
+
 .background {
   background-image: url(img/background.jpg);
-  background-repeat: no-repeat;
+  background-repeat: repeat;
   background-size: cover;
   background-position: center center;
+  font-family: "Shlop", sans-serif;
+}
+
+.box-logo {
+  width: 25vw;
+  display: flex;
+  margin: auto auto 1.5rem;
 }
 
 .box {
-    display: flex;
-    justify-content: center;
-    margin-bottom: 1rem;
+  display: flex;
+  justify-content: center;
+  margin-bottom: 1rem;
 }
 
 .box-image img {
-    width: 40vw;
-    animation: anima;
-    animation-duration: 800ms;
+  width: 40vw;
+  animation: anima;
+  animation-duration: 800ms;
 }
 
 @keyframes anima {
@@ -36,8 +53,8 @@
 }
 
 .bolinhas {
-    justify-content: center;
-    display: flex;
+  justify-content: center;
+  display: flex;
 }
 
 .bolinhas span {
@@ -47,15 +64,15 @@
   margin-right: 1px;
   border-radius: 50%;
   cursor: pointer;
-  background: #85d97aa8;
+  background: var(--light-green);
 }
 
 .bolinhas span.marcador {
-  background-color: #5FE431;
+  background-color: var(--green-neon);
 }
 
 footer {
-  background-color: #11aebf;
+  background-color: var(--light-blue);
   opacity: 60%;
   font-size: large;
   margin-top: 3rem;
@@ -65,4 +82,9 @@ footer {
   display: flex;
   align-items: center;
   justify-content: space-evenly;
+}
+
+footer .criadoras {
+  color: #000000;
+  text-shadow: 1px 1px 2px var(--green-neon);
 }
